# docs/openapi.yaml
openapi: 3.0.1
info:
  title: Bank REST API
  description: RESTful API для управления банковскими картами с JWT аутентификацией
  version: 1.0.0
  contact:
    name: Oleg Kovalev
servers:
  - url: http://localhost:8080
    description: Локальный сервер
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Введите JWT токен в формате "Bearer {token}"
  schemas:
...
    RegisterRequest:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 20
          example: "johndoe"
        email:
          type: string
          format: email
          example: "john@example.com"
        password:
          type: string
          minLength: 6
          maxLength: 40
          example: "password123"
        fullName:
          type: string
          example: "John Doe"
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: "johndoe"
        password:
          type: string
          example: "password123"
...
    JwtResponse:
      type: object
      properties:
        token:
          type: string
          example: "eyJhbGciOiJIUzUxMiJ9..."
        type:
          type: string
          example: "Bearer"
        username:
          type: string
          example: "johndoe"
        email:
          type: string
          example: "john@example.com"
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          example: 400
        error:
          type: string
          example: "Bad Request"
        message:
          type: string
          example: "Invalid input parameters"
        validationErrors:
          type: object
          additionalProperties:
            type: string
...
security:
  - bearerAuth: []
paths:
...
  /auth/register/user:
    post:
      tags:
        - Authentication
      summary: Регистрация обычного пользователя
      description: Создаёт нового пользователя с ролью USER
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Пользователь успешно зарегистрирован
          content:
            text/plain:
              schema:
                type: string
                example: "User registered successfully with roles: [ROLE_USER]"
        '400':
          description: Ошибка валидации или пользователь уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/register/admin:
    post:
      tags:
        - Authentication
      summary: Регистрация администратора
      description: Создаёт нового пользователя с ролями ADMIN и USER (только для ADMIN)
      operationId: registerAdmin
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Администратор успешно зарегистрирован
          content:
            text/plain:
              schema:
                type: string
                example: "User registered successfully with roles: [ROLE_ADMIN, ROLE_USER]"
        '400':
          description: Ошибка валидации или пользователь уже существует
        '403':
          description: Доступ запрещён (не ADMIN)
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Вход в систему
      description: Аутентификация пользователя и получение JWT токена
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Успешная аутентификация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JwtResponse'
        '401':
          description: Неверные учетные данные

  # ========== USER CARD CONTROLLER ==========
...